requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
python-telegram-bot>=22.3
sqlalchemy>=2.0.43
python-dotenv>=1.1.1
pydantic>=2.11.7